Add point factory object
For now there's not much use for the factory object except that one does
not have to use 'new' explicitly. In theory the factory object gives an
opportunity to change the class without breaking the contract with
clients but there's not much variation in constructing a point other
than specifying its coordinates.